refactor: highlight code into HighLightReferences.js
